feat(dotcom): show the invite sign-in dialog on top of the anonymous editor
The invite route now stores the token and resolves to the normal root
experience (the anonymous editor when signed out, your most recent file
when signed in) instead of a blank page. WorkspaceInviteHandler gains a
signed-out branch that fetches the invite info and shows the sign-in
dialog on top, leaving the token in session storage so the existing
signed-in branch completes the join after authentication.
